% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bignumSpace.R
\name{bignumSpace}
\alias{bignumSpace}
\title{Place space between large numbers}
\usage{
bignumSpace(x)
}
\arguments{
\item{x}{is number to apply format}
}
\value{
character version of number
}
\description{
Function separates big numbers by space: like 10 000 000
}
\author{
Jani Miettinen <jani.k.miettinen@gmail.com>
}
