% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyFiles.R
\name{copyFiles}
\alias{copyFiles}
\title{Copy files}
\usage{
copyFiles(source, destination, overwrite = FALSE, file.delete = FALSE)
}
\arguments{
\item{source}{source folder}

\item{destination}{destination folder}

\item{overwrite}{overwrite files from destination folder?}

\item{file.delete}{deletes files from source after copying?}
}
\value{
logical operator
}
\description{
Copies file structure from source to destination.
}
\author{
Jani Miettinen <jani.k.miettinen@gmail.com>
}
