% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateDDMMYYYY.R
\name{dateDDMMYYYY}
\alias{dateDDMMYYYY}
\title{Date transform from DDMMYYY}
\usage{
dateDDMMYYYY(x)
}
\arguments{
\item{x}{is date to transform}
}
\value{
date
}
\description{
Function transforms date from 31DEC1985 to right format
}
\author{
Jani Miettinen <jani.k.miettinen@gmail.com>
}
