% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateLONGFORMAT.R
\name{dataLONGFORMAT}
\alias{dataLONGFORMAT}
\title{Transform "Tue Oct 27 2015" to date}
\usage{
dataLONGFORMAT(x)
}
\arguments{
\item{x}{is variable to transform}
}
\value{
date
}
\description{
Function transforms "Tue Oct 27 2015" typish values to date format.
}
\author{
Jani Miettinen <jani.k.miettinen@gmail.com>
}
