% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funAUTOMATOR.R
\name{funAUTOMATOR}
\alias{funAUTOMATOR}
\title{Fun operator}
\usage{
funAUTOMATOR(x)
}
\value{
"Mä vaan duunaan.."
}
\description{
Function to play around.
}
\author{
Jani Miettinen <jani.k.miettinen@gmail.com>
}
