% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_github.R
\name{install_github}
\alias{install_github}
\title{Install development version from Github}
\usage{
install_github(repo = "rstats-db/odbc", ref = "master", sudo = FALSE)
}
\arguments{
\item{repo}{username and repository name required as in devtools::install_github()}

\item{sudo}{logical TRUE if you are installing to system library. Default is FALSE}
}
\value{
installs a packages
}
\description{
Downloads package source as .tar.gz and installs package
}
\examples{
 \dontrun{
 install_github("rstudio/rmarkdown")
 }
}
\author{
Markus Kainu <markus.kainu@kela.fi>
}
