% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installPKGS.R
\name{installPKGS}
\alias{installPKGS}
\title{Install basic packages}
\usage{
installPKGS(PKGS = "http://muuankarski.kapsi.fi/luntti/r-paketit.txt")
}
\arguments{
\item{PKGS}{is vector of packages}
}
\description{
Installs basic packages to R.
}
\author{
Jani Miettinen <jani.k.miettinen@gmail.com>
}
