% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDatasets.R
\name{loadDatasets}
\alias{loadDatasets}
\title{Load datasets from folder}
\usage{
loadDatasets(folder, format = "rds")
}
\arguments{
\item{format}{file format to load; rds or rdata}
}
\description{
Loads all dataset from folder
}
\author{
Jani Miettinen <jani.k.miettinen@gmail.com>
}
