% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNEWFLEX.R, R/makeNEWSITE.R
\name{makeNEWFLEX}
\alias{makeNEWFLEX}
\alias{makeNEWSITE}
\title{Make new flexdashboard website}
\usage{
makeNEWFLEX(destination = "./", overwrite = FALSE)

makeNEWSITE(destination, overwrite = FALSE)
}
\arguments{
\item{destination}{is folder where to create files.}

\item{overwrite}{overwrite existing files.}

\item{destination}{is folder where to create files.}

\item{overwrite}{overwrite existing files.}
}
\value{
files

files
}
\description{
Function creates flexdasboard template to folder.

Function creates "_site.yml", "index.rmd", "footer.html" and "styles.css" files.
}
\author{
Jani Miettinen <jani.k.miettinen@gmail.com>

Jani Miettinen <jani.k.miettinen@gmail.com>
}
