% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeOLDBCKP.R
\name{makeOLDBCKP}
\alias{makeOLDBCKP}
\title{Backup filesystem}
\usage{
makeOLDBCKP(x)
}
\description{
Creates backup ZIP-file from root-folder. Exludes csv, rds, rdata, sas7bdat and zip.
}
\author{
Jani Miettinen <jani.k.miettinen@gmail.com>
}
