% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_flexdashboard.R
\name{new_flexdashboard}
\alias{new_flexdashboard}
\title{Make new flexdashboard website}
\usage{
new_flexdashboard(destination = "./", overwrite = FALSE)
}
\arguments{
\item{destination}{is folder where to create files.}

\item{overwrite}{overwrite existing files.}
}
\value{
files
}
\description{
Function creates flexdasboard template to folder.
}
\author{
Jani Miettinen <jani.k.miettinen@gmail.com>
}
