% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfibo.R
\name{testfibo}
\alias{testfibo}
\title{Fibonacci Benchmark test}
\usage{
testfibo(n = 1e+05)
}
\arguments{
\item{n}{number of Fibonacci values}
}
\value{
string of benchmark results
}
\description{
Function test prosessor speed by calculating Fibonacci numbers.
}
\examples{
 \dontrun{
 for(i in 1:10)testfibo(n=200000)
 }
}
\author{
Jani Miettinen <jani.k.miettinen@gmail.com>
}
