% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNEWSITE.R
\name{makeNEWSITE}
\alias{makeNEWSITE}
\title{Make new website}
\usage{
makeNEWSITE(destination, overwrite = FALSE)
}
\arguments{
\item{destination}{is folder where to create files.}

\item{overwrite}{overwrite existing files.}
}
\value{
files
}
\description{
Function creates "_site.yml", "index.rmd", "footer.html" and "styles.css" files.
}
\author{
Jani Miettinen <jani.k.miettinen@gmail.com>
}
